+++ src/test/java/com/ays/AbstractRestControllerTest.java
@@ -10,7 +10,6 @@ import com.ays.parameter.model.AysParameterBuilder;
 import com.ays.parameter.service.AysParameterService;
 import com.ays.super_admin.entity.SuperAdminEntityBuilder;
 import com.ays.user.model.entity.UserEntityBuilder;
-import com.ays.util.AysMockMvc;
 import io.jsonwebtoken.JwtBuilder;
 import io.jsonwebtoken.Jwts;
 import org.apache.commons.lang3.time.DateUtils;
